+++ packages/plugin-packages/mashroom-security-provider-openid-connect/src/callback/mashroom-bootstrap.ts
@@ -1,0 +1,13 @@
+import express from 'express';
+import {MashroomApiPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
+import callbackRoute from './mashroom-openid-connect-callback-route';
+
+const bootstrap: MashroomApiPluginBootstrapFunction = async (pluginName, pluginConfig, contextHolder) => {
+    const defaultBackUrl = contextHolder.getPluginContext().serverConfig.indexPage;
+    const router = express.Router();
+    // @ts-ignore
+    router.all('/*', callbackRoute(defaultBackUrl));
+    return router;
+};
+
+export default bootstrap;